chore: migrate from Askama to Rinja
Askama was deprecated in favor of Rinja (a fork of Askama), so this gets
us the latest features available (Oct 2024 release instead of May 2024).

However, Rinja will soon be deprecated in favor of Askama (yes, really),
so when released it'll be necessary to migrate back. Nevertheless, the
new Askama will have Rinja's current contents, so this commit is still
important to the upgrade, especially because it switches from using the
`askama_rocket` dependency (which no longer exists) to directly
rendering templates. (There is no `rinja_rocket` equivalent. (Well yes
there is but it was deprecated; they want us to do it manually instead
of integrating with Rocket themselves.))

Note that the link below refers to the original, "old" Askama repository
before Rinja's is renamed to Askama, so in the future it might require
manual adjustment to whatever repository name the old archive is
assigned.

See: https://github.com/rinja-rs/askama/issues/1125