feat: update permission management requirements
This makes it so users are only allowed to manage a system's Permissions
if they have permission to *both* manage the system and manage the
system's permissions. For just listing, having either continues to be
sufficient.

This is useful e.g. when someone should be able to assign permissions,
but not delete them altogether. In the future, it might make sense to
consider having a separate `$hive:assign-perms:<scope>` permission and
reverting this commit so that `$hive:manage-perms:<scope>` refers
exclusively to literal permission management.