feat: require user when PermsEvaluator is built
Previously, requiring the PermsEvaluator request guard allows no user
to be logged in, with any further permission queries failing as "not
allowed", but that did not allow for distinguishing the case of a user
needing to log in to access more resources.

This commit makes it so simply constructing a PermsEvaluator instance is
only possible if there is any user logged in. Routes and etc. can still
request an Option<PermsEvaluator> if the previous behavior is needed,
but this makes everything more explicit.