fix: remove unused extraneous `at` param from all_subgroups_of function
When `all_subgroups_of` was originally created, it was accidentally
given an extra `at DATE` parameter, which does not make sense since
subgroups are not temporally-bound like direct memberships are.

Technically, this commit should create a new migration to drop both
`all_subgroups_of` and `all_members_of` (which calls it), but that would
just be confusing and bring no added benefit since this code is not
currently deployed anywhere anyway.