feat: redirect to permission details on creation
This removes the separate "permission created successfully" page and
just redirects to the details page instead so it can be assigned right
away. Keeping the partial version would also not be very beneficial
since it completely replaces the creation form anyway, so the user would
have to refresh the page to continue adding more permissions.