fix: use relative path to ./static directory
This changes the path to the `./static` directory so that it is
evaluated only at runtime based on the current working directory.

Previously, the `relative!` macro made it so that the absolute path was
hardcoded at build time based on the crate root directory during
compilation, which works for most use cases (including local and Docker
with identical `WORKDIR`s), but should not really be an assumption for
us to make (build-time absolute path == runtime absolute path).