fix: don't resolve username on tag unassign
This was broken because it `unwrap`'d assuming a username must always
exist, but this is not true if the assignment was for a group. In any
case, the operation itself is unnecessary since the label would never be
read in this scenario, so we might as well avoid it.